Use GradleTestExtension to provide Project and apply plugins
